requirements.txt
neo4j==5.12.0 
pandas==2.1.0
py4j
pyarrow==13.0.0
psycopg2-binary~=2.9.6
fastparquet==2023.7.0